requirements.txt
# Basics
ruamel.yaml
plyfile
networkx
numpy
metis
psutil

# SciKits
scikit-image
scikit-learn
